comments機能実装、コメント数、削除

+++ app/Http/Controllers/CommentController.php
@@ -29,6 +29,10 @@ class CommentController extends Controller
 
     public function destroy(Comment $comment)
     {
+        if(auth()->id() !== $comment->user_id) {
+            abort(403);
+        }
+
         $comment->delete();
 
         return back()->with('status', 'コメントを削除しました。');
